requirements.txt
web3>=7.0
py-solc-x>=2.0
textual>=0.50
